image_file_view exposed and merged master
