+++ lib/src/image_file_view/io_preview.dart
@@ -4,8 +4,9 @@ import '../image_file.dart';
 
 class ImageFileView extends StatelessWidget {
   final ImageFile file;
+  final BoxFit fit;
 
-  const ImageFileView({Key? key, required this.file}) : super(key: key);
+  const ImageFileView({super.key, required this.file, this.fit = BoxFit.cover});
 
   @override
   Widget build(BuildContext context) {
@@ -13,7 +14,7 @@ class ImageFileView extends StatelessWidget {
       borderRadius: BorderRadius.circular(4),
       child: Image.file(
         File(file.path!),
-        fit: BoxFit.cover,
+        fit: fit,
       ),
     );
   }
