+++ lib/src/multi_image_picker_view.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_reorderable_grid_view/entities/order_update_entity.dart';
 import 'package:flutter_reorderable_grid_view/widgets/reorderable_builder.dart';
 
-import 'preview_item.dart';
+import 'list_image_item.dart';
 
 import '../multi_image_picker_view.dart';
